+++ src/loader/replace-template-tag/types.ts
@@ -1,0 +1,8 @@
+export type IReplaceTemplateTagOptions = {
+  replaceTmpTagMap?: {
+    [k: string]: {
+      mp: string;
+      web: string;
+    }
+  }
+};