+++ src/plugin/rem-to-rpx/rem-to-rpx.ts
@@ -1,4 +1,4 @@
-import { handleRem } from './css-handler';
+import { transFormRem } from 't-comm/lib/rem/rem';
 import { PLATFORM_MAP } from '../../helper/config';
 
 function isInWhiteList(whiteList, fileName) {
@@ -59,7 +59,7 @@ export class RemToRpxPlugin {
         const sourceCode = asset.source() || asset._valueAsString || asset._value || asset._cachedSource;
         if (sourceCode != null) {
           // 这里返回null或者undefined会导致编译过程无法结束，所以sourceCode需要判空才给asset.source赋值
-          const newSource = handleRem(sourceCode, this.factor, this.unit);
+          const newSource = transFormRem(sourceCode, this.factor, this.unit);
 
           compilation.assets[fileName] = {
             source() {